refactor: layout and css
